fix(Table): fix the calculation error of cell height when using renderRowExpanded
